Got basic position movement working again
