+++ src/sample/fluid2D/densityWGSL.ts
